requirements.txt
numpy==1.24.3
opencv-python==4.7.0.72
onnx==1.14.0
onnxruntime-gpu==1.15.0
insightface==0.7.3
psutil==5.9.5
tk==0.1.0
torch==2.0.1